City name working!
Finally got the first part of the backend done. Now the API and front end communicate through a POST request, getting the name of the city and using openmeteo's geolocation API to acquire latitude and longitude information of the given city. Now, I have to make the backend for the actual forecast information and display it on the front end. Got a loong way to go.